+++ src/main/java/com/fixadate/domain/adate/repository/AdateJpaRepository.java
@@ -1,7 +1,5 @@
 package com.fixadate.domain.adate.repository;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,5 +8,4 @@ import com.fixadate.domain.adate.entity.Adate;
 @Repository
 public interface AdateJpaRepository extends JpaRepository<Adate, Long> {
 
-	Optional<Adate> findAdateByCalendarId(final String calendarId);
 }
