+++ src/test/java/com/fixadate/integration/domain/googleservice/GoogleServiceTest.java
@@ -25,7 +25,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 
 import com.fixadate.config.DataClearExtension;
 import com.fixadate.domain.adate.entity.Adate;
-import com.fixadate.domain.adate.repository.AdateJpaRepository;
+import com.fixadate.domain.adate.service.repository.AdateRepository;
 import com.fixadate.domain.auth.dto.request.MemberRegisterRequest;
 import com.fixadate.domain.auth.service.AuthService;
 import com.fixadate.domain.googlecalendar.service.GoogleService;
@@ -49,7 +49,7 @@ public class GoogleServiceTest {
 	@Autowired
 	private MemberRepository memberRepository;
 	@Autowired
-	private AdateJpaRepository adateJpaRepository;
+	private AdateRepository adateRepository;
 	@Autowired
 	private AuthService authService;
 
@@ -80,7 +80,7 @@ public class GoogleServiceTest {
 			() -> assertDoesNotThrow(() -> googleService.syncEvents(events, member.get())));
 
 		for (Adate adate : adates) {
-			assertNotNull(adateJpaRepository.findAdateByCalendarId(adate.getCalendarId()));
+			assertNotNull(adateRepository.findAdateByCalendarId(adate.getCalendarId()));
 		}
 	}
